+++ src/tools/DeepAnalysis/FnVar/TFnVarDef.ts
@@ -8,5 +8,5 @@ export type TGetFnDefNeed = {
     paramMap: TParamMapIn;
     GValMap: TGValMap;
     lStrTrimLen: number;
-    comment: string;
+    lineComment: string;
 };