+++ src/tools/DeepAnalysis/FnVar/def/varSetCapacityFunc.ts
@@ -10,7 +10,7 @@ export function varSetCapacityFunc({
     paramMap,
     GValMap,
     lStrTrimLen,
-    comment,
+    lineComment,
 }: TGetFnDefNeed): void {
     // eslint-disable-next-line no-magic-numbers
     if (lStrTrimLen < 8) return; // 'NumGet('.length
@@ -25,7 +25,7 @@ export function varSetCapacityFunc({
 
         const character: number = ch + v[0].indexOf(RawName);
 
-        const value: TValMetaIn = getValMeta(line, character, RawName, valMap, comment);
+        const value: TValMetaIn = getValMeta(line, character, RawName, valMap, lineComment);
         valMap.set(RawName.toUpperCase(), value);
     }
 }